requirements.txt
accelerate==1.8.1
albucore==0.0.24
albumentations==2.0.8
alpha-clip==1.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
beautifulsoup4==4.13.4
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
clip==1.0
cloudpickle==3.1.1
contourpy==1.3.2
cycler==0.12.1
diffusers==0.34.0
easydict==1.13
einops==0.7.0
eval_type_backport==0.2.2
exceptiongroup==1.3.0
Farama-Notifications==0.0.4
filelock==3.18.0
fonttools==4.58.5
fsspec==2025.5.1
ftfy==6.3.1
gdown==5.2.0
gymnasium==1.1.1
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.2
idna==3.10
imageio==2.37.0
importlib_metadata==8.7.0
jaxtyping==0.2.31
Jinja2==3.1.6
kiwisolver==1.4.8
kornia==0.8.1
kornia_rs==0.1.9
lazy_loader==0.4
llvmlite==0.44.0
loralib==0.1.2
MarkupSafe==3.0.2
matplotlib==3.10.3
more-itertools==10.7.0
mpmath==1.3.0
networkx==3.4.2
numba==0.61.2
numpy==1.26.4
oauthlib==3.3.1
omegaconf==2.3.0
openai-whisper==20250625
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
packaging==25.0
pandas==2.3.1
pillow==11.3.0
psutil==7.0.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pyglet==2.1.6
PyMatting==1.1.14
pyparsing==3.2.3
pyqtgraph==0.13.7
PySide6==6.9.1
PySide6_Addons==6.9.1
PySide6_Essentials==6.9.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
repath==0.9.0
requests==2.32.4
safetensors==0.5.3
scikit-image==0.25.2
scipy==1.15.3
shiboken6==6.9.1
simsimd==6.5.0
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.2
soupsieve==2.7
stable_baselines3==2.6.0
stringzilla==3.12.5
sympy==1.14.0
texture-baker==0.0.1
tifffile==2025.5.10
tiktoken==0.9.0
timm==1.0.16
tokenizers==0.19.1
torch==2.7.1
torchaudio==2.7.1
torchvision==0.22.1
tqdm==4.67.1
transformers==4.42.3
transparent-background==1.3.4
trimesh==4.4.1
typeguard==2.13.3
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uv-unwrapper==0.0.1
wadler_lindig==0.1.7
wcwidth==0.2.13
wget==3.2
zipp==3.23.0
